Requiring an HTTP Header token for sending
