+++ src/main/resources/application.properties
@@ -32,6 +32,9 @@ phase4.endpoint.address=http://localhost:8080/as4
 # [CHANGEME] Public URL of your SMP to check for valid inbound requests
 #smp.url=http://smp.helger.com
 
+# [CHANGEME] The mandatory value required in the "X-Token" HTTP header for the sending APIs
+phase4.api.requiredtoken=NjIh9tIx3Rgzme19mGIy
+
 # [CHANGEME] put your keystore here
 org.apache.wss4j.crypto.merlin.keystore.type=JKS
 org.apache.wss4j.crypto.merlin.keystore.file=invalid-keystore-pw-peppol.jks
